extract generic base and add initial mainnet subgraph
