+++ packages/subgraph-base/src/gno.ts
@@ -5,7 +5,7 @@ import {
   store,
   ethereum,
 } from "@graphprotocol/graph-ts";
-import { Transfer } from "../generated/ds-gno/GNO";
+import { Transfer } from "../generated/ds-gno/ERC20";
 import { loadOrCreateUser, ADDRESS_ZERO, removeOrSaveUser } from "./helpers";
 
 export function handleTransfer(event: Transfer): void {
