+++ packages/subgraph-base/subgraph.yaml
@@ -2,12 +2,12 @@ specVersion: 0.0.4
 schema:
   file: ./schema.graphql
 dataSources:
- - kind: ethereum
+  - kind: ethereum
     name: ds-gno
     network: mainnet
     source:
       address: "0x0000000000000000000000000000000000000000"
-      abi: GNO
+      abi: ERC20
       startBlock: 0
     mapping:
       kind: ethereum/events
@@ -27,7 +27,7 @@ dataSources:
     network: mainnet
     source:
       address: "0x0000000000000000000000000000000000000000"
-      abi: LGNO
+      abi: ERC20
       startBlock: 0
     mapping:
       kind: ethereum/events
@@ -59,8 +59,6 @@ dataSources:
       abis:
         - name: Factory
           file: ./abis/factory.json
-        - name: ERC20
-          file: ./abis/ERC20.json
       eventHandlers:
         - event: PairCreated(indexed address,indexed address,address,uint256)
           handler: handleNewPair
@@ -80,8 +78,6 @@ templates:
       abis:
         - name: Pair
           file: ./abis/pair.json
-        - name: Factory
-          file: ./abis/factory.json
         - name: ERC20
           file: ./abis/ERC20.json
       eventHandlers:
