+++ packages/subgraph-gnosis-chain/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "subgraph-voting-power",
+  "name": "subgraph-gnosis-chain",
   "license": "UNLICENSED",
   "author": "Cristóvão Honorato",
   "scripts": {
