+++ packages/subgraph-gnosis-chain/subgraph.yaml
@@ -7,7 +7,7 @@ dataSources:
     network: xdai
     source:
       address: "0x9C58BAcC331c9aa871AFD802DB6379a98e80CEdb"
-      abi: GNO
+      abi: ERC20
       startBlock: 11629829
     mapping:
       kind: ethereum/events
@@ -16,18 +16,18 @@ dataSources:
       entities:
         - User
       abis:
-        - name: GNO
-          file: ./abis/gno.json
+        - name: ERC20
+          file: ../subgraph-base/abis/ERC20.json
       eventHandlers:
         - event: Transfer(indexed address,indexed address,uint256)
           handler: handleTransfer
-      file: ./src/gno.ts
+      file: ../subgraph-base/src/gno.ts
   - kind: ethereum
     name: ds-mgno
     network: xdai
     source:
       address: "0x722fc4DAABFEaff81b97894fC623f91814a1BF68"
-      abi: MGNO
+      abi: ERC20
       startBlock: 19468779
     mapping:
       kind: ethereum/events
@@ -36,8 +36,8 @@ dataSources:
       entities:
         - User
       abis:
-        - name: MGNO
-          file: ./abis/SBCToken.json
+        - name: ERC20
+          file: ../subgraph-base/abis/ERC20.json
       eventHandlers:
         - event: Transfer(indexed address,indexed address,uint256)
           handler: handleTransfer
@@ -47,7 +47,7 @@ dataSources:
     network: xdai
     source:
       address: "0xd4Ca39f78Bf14BfaB75226AC833b1858dB16f9a1"
-      abi: LGNO
+      abi: ERC20
       startBlock: 20388099
     mapping:
       kind: ethereum/events
@@ -56,12 +56,12 @@ dataSources:
       entities:
         - User
       abis:
-        - name: LGNO
-          file: ./abis/lgno.json
+        - name: ERC20
+          file: ../subgraph-base/abis/ERC20.json
       eventHandlers:
         - event: Transfer(indexed address,indexed address,uint256)
           handler: handleTransfer
-      file: ./src/lgno.ts
+      file: ../subgraph-base/src/lgno.ts
   - kind: ethereum/contract
     name: HoneySwap
     network: xdai
@@ -73,14 +73,12 @@ dataSources:
       kind: ethereum/events
       apiVersion: 0.0.6
       language: wasm/assemblyscript
-      file: ./src/HoneySwap.ts
+      file: ../subgraph-base/src/factory.ts
       entities:
         - User
       abis:
         - name: Factory
-          file: ./abis/factory.json
-        - name: ERC20
-          file: ./abis/ERC20.json
+          file: ../subgraph-base/abis/factory.json
       eventHandlers:
         - event: PairCreated(indexed address,indexed address,address,uint256)
           handler: handleNewPair
@@ -95,14 +93,12 @@ dataSources:
       kind: ethereum/events
       apiVersion: 0.0.6
       language: wasm/assemblyscript
-      file: ./src/Swapr.ts
+      file: ../subgraph-base/src/factory.ts
       entities:
         - User
       abis:
         - name: Factory
-          file: ./abis/factory.json
-        - name: ERC20
-          file: ./abis/ERC20.json
+          file: ../subgraph-base/abis/factory.json
       eventHandlers:
         - event: PairCreated(indexed address,indexed address,address,uint256)
           handler: handleNewPair
@@ -117,14 +113,12 @@ dataSources:
       kind: ethereum/events
       apiVersion: 0.0.6
       language: wasm/assemblyscript
-      file: ./src/Sushi.ts
+      file: ../subgraph-base/src/factory.ts
       entities:
         - User
       abis:
         - name: Factory
-          file: ./abis/factory.json
-        - name: ERC20
-          file: ./abis/ERC20.json
+          file: ../subgraph-base/abis/factory.json
       eventHandlers:
         - event: PairCreated(indexed address,indexed address,address,uint256)
           handler: handleNewPair
@@ -139,14 +133,12 @@ dataSources:
       kind: ethereum/events
       apiVersion: 0.0.6
       language: wasm/assemblyscript
-      file: ./src/Elk.ts
+      file: ../subgraph-base/src/factory.ts
       entities:
         - User
       abis:
         - name: Factory
-          file: ./abis/factory.json
-        - name: ERC20
-          file: ./abis/ERC20.json
+          file: ../subgraph-base/abis/factory.json
       eventHandlers:
         - event: PairCreated(indexed address,indexed address,address,uint256)
           handler: handleNewPair
@@ -161,14 +153,12 @@ dataSources:
       kind: ethereum/events
       apiVersion: 0.0.6
       language: wasm/assemblyscript
-      file: ./src/BaoSwap.ts
+      file: ../subgraph-base/src/factory.ts
       entities:
         - User
       abis:
         - name: Factory
-          file: ./abis/factory.json
-        - name: ERC20
-          file: ./abis/ERC20.json
+          file: ../subgraph-base/abis/factory.json
       eventHandlers:
         - event: PairCreated(indexed address,indexed address,address,uint256)
           handler: handleNewPair
@@ -183,14 +173,12 @@ dataSources:
       kind: ethereum/events
       apiVersion: 0.0.6
       language: wasm/assemblyscript
-      file: ./src/Symmetric.ts
+      file: ../subgraph-base/src/factory.ts
       entities:
         - User
       abis:
         - name: Factory
-          file: ./abis/factory.json
-        - name: ERC20
-          file: ./abis/ERC20.json
+          file: ../subgraph-base/abis/factory.json
       eventHandlers:
         - event: PairCreated(indexed address,indexed address,address,uint256)
           handler: handleNewPair
@@ -204,16 +192,14 @@ templates:
       kind: ethereum/events
       apiVersion: 0.0.6
       language: wasm/assemblyscript
-      file: ./src/pair.ts
+      file: ../subgraph-base/src/pair.ts
       entities:
         - AMMPair
       abis:
         - name: Pair
-          file: ./abis/pair.json
-        - name: Factory
-          file: ./abis/factory.json
+          file: ../subgraph-base/abis/pair.json
         - name: ERC20
-          file: ./abis/ERC20.json
+          file: ../subgraph-base/abis/ERC20.json
       eventHandlers:
         - event: Transfer(indexed address,indexed address,uint256)
           handler: handleTransfer
