+++ packages/subgraph-gnosis-chain/tests/helpers.ts
@@ -1,115 +1,20 @@
-import {
-  createMockedFunction,
-  clearStore,
-  test,
-  assert,
-  logStore,
-} from "matchstick-as/assembly/index";
-import { Address, BigInt, Bytes, ethereum } from "@graphprotocol/graph-ts";
-import { User } from "../generated/schema";
-import { handleTransfer } from "../src/mgno";
-import { log, newMockEvent } from "matchstick-as";
-import { ADDRESS_ZERO } from "../src/helpers";
-import { Transfer } from "../generated/templates/Pair/ERC20";
-import { Sync } from "../generated/templates/Pair/Pair";
-import { PairCreated } from "../generated/HoneySwap/Factory";
+import { BigInt, Address } from "@graphprotocol/graph-ts";
 
-export function createTransferEvent(
-  from: Address,
-  to: Address,
-  value: BigInt,
-  data: string
-): Transfer {
-  let mockEvent = newMockEvent();
+export const DEPOSIT_ADDRESS = Address.fromString(
+  "0x0B98057eA310F4d31F2a452B414647007d1645d9"
+);
 
-  mockEvent.parameters = new Array();
+export const ADDRESS_ZERO = Address.fromString(
+  "0x0000000000000000000000000000000000000000"
+);
+export const USER1_ADDRESS = Address.fromString(
+  "0x0000000000000000000000000000000000000001"
+);
+export const USER2_ADDRESS = Address.fromString(
+  "0x0000000000000000000000000000000000000002"
+);
 
-  mockEvent.parameters.push(
-    new ethereum.EventParam("from", ethereum.Value.fromAddress(from))
-  );
-  mockEvent.parameters.push(
-    new ethereum.EventParam("to", ethereum.Value.fromAddress(to))
-  );
-  mockEvent.parameters.push(
-    new ethereum.EventParam("value", ethereum.Value.fromSignedBigInt(value))
-  );
-  mockEvent.parameters.push(
-    new ethereum.EventParam("data", ethereum.Value.fromString(data))
-  );
+export const data = "0x00";
 
-  let newTransferEvent = new Transfer(
-    mockEvent.address,
-    mockEvent.logIndex,
-    mockEvent.transactionLogIndex,
-    mockEvent.logType,
-    mockEvent.block,
-    mockEvent.transaction,
-    mockEvent.parameters
-  );
-
-  return newTransferEvent;
-}
-
-export function createSyncEvent(reserve0: BigInt, reserve1: BigInt): Sync {
-  let mockEvent = newMockEvent();
-
-  mockEvent.parameters = new Array();
-
-  mockEvent.parameters.push(
-    new ethereum.EventParam("value", ethereum.Value.fromSignedBigInt(reserve0))
-  );
-  mockEvent.parameters.push(
-    new ethereum.EventParam("value", ethereum.Value.fromSignedBigInt(reserve1))
-  );
-
-  let newSyncEvent = new Sync(
-    mockEvent.address,
-    mockEvent.logIndex,
-    mockEvent.transactionLogIndex,
-    mockEvent.logType,
-    mockEvent.block,
-    mockEvent.transaction,
-    mockEvent.parameters
-  );
-
-  return newSyncEvent;
-}
-
-export function createPairCreatedEvent(
-  token0: Address,
-  token1: Address,
-  pair: Address,
-  value: BigInt
-): PairCreated {
-  let mockEvent = newMockEvent();
-
-  mockEvent.parameters = new Array();
-
-  mockEvent.parameters.push(
-    new ethereum.EventParam("token0", ethereum.Value.fromAddress(token0))
-  );
-
-  mockEvent.parameters.push(
-    new ethereum.EventParam("token1", ethereum.Value.fromAddress(token1))
-  );
-
-  mockEvent.parameters.push(
-    new ethereum.EventParam("pair", ethereum.Value.fromAddress(pair))
-  );
-
-  mockEvent.parameters.push(
-    new ethereum.EventParam("", ethereum.Value.fromSignedBigInt(value))
-  );
-
-  let newPairCreatedEvent = new PairCreated(
-    mockEvent.address,
-    mockEvent.logIndex,
-    mockEvent.transactionLogIndex,
-    mockEvent.logType,
-    mockEvent.block,
-    mockEvent.transaction,
-    mockEvent.parameters
-  );
-
-  return newPairCreatedEvent;
-}
+export const MGNO_PER_GNO = BigInt.fromString("32");
+export const ONE_GNO = BigInt.fromString("1000000000000000000");
