+++ packages/subgraph-mainnet/package.json
@@ -1,9 +1,10 @@
 {
-  "name": "gno-voting-power-eth",
+  "name": "subgraph-mainnet",
   "license": "UNLICENSED",
   "scripts": {
-    "codegen": "rm -rf generated && graph codegen",
-    "build": "graph build",
+    "doall": "rm -rf generated && rm -rf build && graph codegen && graph build",
+    "codegen": "graph codegen",
+    "test": "graph test",
     "deploy": "graph deploy --node https://api.thegraph.com/deploy/ jfschwarz/gno-voting-power-eth",
     "create-local": "graph create --node http://localhost:8020/ jfschwarz/gno-voting-power-eth",
     "remove-local": "graph remove --node http://localhost:8020/ jfschwarz/gno-voting-power-eth",
